fix(editor3): Refactor deleteBlock
